Renamd the barcode scanner file.
